requirements.txt
accelerate==1.11.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.3.0
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langchain==1.0.2
langchain-core==1.0.1
langgraph==1.0.1
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.9
langsmith==0.4.38
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
narwhals==2.10.0
networkx==3.4.2
numpy==2.2.6
openai==2.6.1
orjson==3.11.4
ormsgpack==1.11.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==6.33.0
psutil==7.1.3
pyarrow==22.0.0
pydantic==2.12.3
pydantic_core==2.41.4
pydeck==0.9.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
reportlab==4.4.4
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.28.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
sentencepiece==0.2.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
xxhash==3.6.0
zstandard==0.25.0
